// src/types/pokemon.ts
// Pokemon dto took from GH
// TODO: remove unused fields after architecture is defined
export interface Pokemon {
  id: number
  name: string
  names: {
    name: string
    language: {
      name: string
      url: string
    }
  }[]
  order: number
  height: number
  sprites: {
    front_default: string | null
    other: {
      dream_world: {
        front_default: string | null
      }
      'official-artwork': {
        front_default: string | null
      }
    }
  }
  abilities: {
    ability: {
      name: string
      url: string
    }
    is_hidden: boolean
    slot: number
  }[]
  game_indices: {
    game_index: number
    version: {
      name: string
      url: string
    }
  }[]

  held_items: {
    item: {
      name: string
      url: string
    }
    version_details: {
      rarity: number
      version: {
        name: string
        url: string
      }
    }[]
  }[]
  flavor_text_entries: {
    flavor_text: string
    language: {
      name: string
      url: string
    }
    version: {
      name: string
      url: string
    }
  }[]
  moves: {
    move: {
      name: string
      url: string
    }
    version_group_details: {
      level_learned_at: number
      move_learn_method: {
        name: string
        url: string
      }
      version_group: {
        name: string
        url: string
      }
    }[]
  }[]
  species: {
    name: string
    url: string
  }
  stats: {
    base_stat: number
    effort: number
    stat: {
      name: string
      url: string
    }
  }[]
  types: {
    slot: number
    type: {
      name: string
      url: string
    }
  }[]
  weight: number
  cries?: {
    latest: string
    legacy: string
  }
  items?: {
    name: string

    names: {
      language: {
        name: string
        url: string
      }
      name: string
    }[]
    sprites: {
      default: string | null
    }
  }
}